+++ R/dfm_group.R
@@ -140,12 +140,16 @@ group_dfm <- function(x, features = NULL, documents = NULL, fill = FALSE) {
 
 # select docvar fields that have all the same values within groups
 group_docvars <- function(x, group) {
-    result <- x[match(levels(group), group), sapply(x, is_grouped, group), drop = FALSE]
+    result <- x[match(levels(group), group), sapply(x, is_grouped, as.integer(group)), drop = FALSE]
     rownames(result) <- as.character(levels(group))
     return(result)
 }
 
-# check if there is not within group variance
+# check if values are uniform within groups
 is_grouped <- function(x, group) {
-    all(sapply(split(x, group), function(x) all(x[1] == x)), na.rm = TRUE)
+    if (is.character(x)) {
+        qatd_cpp_is_grouped_character(x, group)
+    } else {
+        qatd_cpp_is_grouped_numeric(as.numeric(x), group)
+    }
 }